requirements.txt
pathway
pandas
requests
datetime
llm_app
python-dotenv